scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pillow==10.1.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.0.3
